+++ lib/widgets/reservation_card.dart
@@ -4,12 +4,16 @@ import '../models/_reservation.dart';
 class ReservationCard extends StatelessWidget {
   final Reservation reservation;
   final int index;
+  final Widget? trailing;
+
+
 
 
 
   const ReservationCard({
     required this.reservation,
     required this.index,
+    this.trailing,
     super.key,
   });
 
@@ -66,6 +70,7 @@ class ReservationCard extends StatelessWidget {
             ],
           ),
         ),
+        trailing: trailing,
       ),
     );
   }